Extract utility functions and bargraph logic
