+++ views/pages/form.ejs
@@ -35,5 +35,3 @@
     <script type="text/javascript" src="js/form.js"></script>
 </body>
 </html>
-
-<!-- action="/" method="POST" enctype="application/x-www-form-urlencoded"  -->
