+++ views/partials/styles/bar-styles.ejs
@@ -19,6 +19,7 @@ body {
 }
 
 .graph div {
+    background-color: attr(data-color);
     position: relative;
     margin-left: 2rem;
 }
